flask-api/requirements.txt
flask
jsonify
clarifai
clarifai_grpc
python_dotenv
